+++ aboutus-mng/src/main/java/cn/springmvc/dao/FriendshipLinkListDao.java
@@ -16,7 +16,7 @@ public interface FriendshipLinkListDao {
 	/**
 	 * 查询该名称的友情链接是否存在
 	 * @param entity 友情链接实体
-	 * @return  friendshipUnitEntity 友情链接实体
+	 * @return  FriendshipUnitEntity 友情链接实体
 	 */
 	public  FriendshipUnitEntity  selectFriendshipLinkIsExistByNAme(FriendshipUnitEntity entity);
 	/**
@@ -29,9 +29,9 @@ public interface FriendshipLinkListDao {
 	/**
 	 * 根据id查询友情链接
 	 * @param id 友情链接ID
-	 * @return  PartnersEntity  友情链接实体
+	 * @return  FriendshipUnitEntity  友情链接实体
 	 */
-	public  FriendshipUnitEntity  selectFriendshipLinkByID(int id);
+	public  FriendshipUnitEntity  selectFriendshipLinkByID(long id);
 	/**
 	 * 前台查询友情链接 
 	 * @return
