+++ aboutus-mng/src/main/java/cn/springmvc/dao/impl/ExpertTeamDaoImpl.java
@@ -1,9 +1,5 @@
 package cn.springmvc.dao.impl;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
+ 
 import javax.annotation.Resource;
 
 import org.apache.ibatis.session.SqlSessionFactory;
@@ -46,7 +42,7 @@ public class ExpertTeamDaoImpl extends SqlSessionDaoSupport  implements ExpertTe
 	}
 
 	@Override
-	public int deleteExpertTeamByID(int id) {
+	public int deleteExpertTeamByID(long id) {
 		
 		int result=0;
 		result = getSqlSession().delete("expertTeam.deleteExpertTeamByID", id);
